requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.44
python-multipart==0.0.12
Jinja2==3.1.4
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.6.1
itsdangerous==2.2.0
python-slugify==8.0.4
pytest==8.3.3
httpx==0.27.2
